# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ZioHelp Helpdesk API",
    "description": "API documentation for ZioHelp Helpdesk system.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "summary": "User registration",
        "description": "Register a new user. Admin approval required before login.",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Email already in use or registration failed"
          },
          "200": {
            "description": "Registration successful. Awaiting admin approval."
          }
        }
      }
    },
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "summary": "User login",
        "description": "Authenticate user and return JWT token and user details.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid credentials or input"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/auth/login-debug": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "loginDebug",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/verify-email": {
      "get": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "verifyEmail",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "email": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}